# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sakura Beauty E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/addresses:
    get:
      operationId: listAddresses
      tags: [users]
      summary: Get user's saved addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [users]
      summary: Add a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressBody"
      responses:
        "201":
          description: Added address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /users/me/addresses/{id}:
    put:
      operationId: updateAddress
      tags: [users]
      summary: Update an address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressBody"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [users]
      summary: Delete an address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        isBlocked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - firstName
        - lastName
        - phone
        - role
        - isBlocked
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - fullName
        - phone
        - street
        - city
        - district
    Address:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - id
        - userId
        - fullName
        - phone
        - street
        - city
        - district
        - postalCode
        - isDefault
...